+++ bootstrap/providers.php
@@ -4,10 +4,12 @@ use App\Providers\AppServiceProvider;
 use App\Providers\Filament\AdminPanelProvider;
 use App\Providers\FortifyServiceProvider;
 use App\Providers\HealthServiceProvider;
+use App\Providers\TelescopeServiceProvider;
 
 return [
     AppServiceProvider::class,
     AdminPanelProvider::class,
     FortifyServiceProvider::class,
     HealthServiceProvider::class,
+    TelescopeServiceProvider::class,
 ];